refactor: make account settings modal more readable and overall improvements
