--- frontend/src/container/CloudIntegrationPage/HeroSection/AccountActions.style.scss
@@ -1,79 +1,0 @@
-.hero-section__actions {
-	margin-top: 12px;
-
-	&-with-account {
-		display: flex;
-		flex-direction: column;
-		gap: 10px;
-	}
-}
-.hero-section__action-buttons {
-	display: flex;
-	align-items: center;
-	gap: 8px;
-}
-
-.hero-section__action-button {
-	font-family: 'Inter';
-	border-radius: 2px;
-	cursor: pointer;
-	font-size: 12px;
-	font-weight: 500;
-	line-height: 16px;
-	padding: 8px 17px;
-
-	&.primary {
-		background: var(--bg-robin-500);
-		border: none;
-		color: var(--bg-vanilla-100);
-	}
-
-	&.secondary {
-		display: flex;
-		align-items: center;
-		border: 1px solid var(--bg-ink-300);
-		color: var(--bg-vanilla-100);
-		border-radius: 2px;
-		background: var(--bg-slate-400);
-		box-shadow: none;
-	}
-}
-
-.cloud-account-selector {
-	border-radius: 2px;
-	border: 1px solid var(--bg-ink-300);
-	background: linear-gradient(
-		139deg,
-		rgba(18, 19, 23, 0.8) 0%,
-		rgba(18, 19, 23, 0.9) 98.68%
-	);
-	box-shadow: 4px 10px 16px 2px rgba(0, 0, 0, 0.2);
-	-webkit-backdrop-filter: blur(20px);
-	backdrop-filter: blur(20px);
-	.ant-select-selector {
-		border-color: var(--bg-slate-400) !important;
-		background: var(--bg-ink-300) !important;
-		padding: 6px 8px !important;
-	}
-	.ant-select-selection-item {
-		color: var(--bg-vanilla-400);
-		font-size: 12px;
-		font-weight: 400;
-		line-height: 16px;
-	}
-
-	.account-option-item {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		&__selected {
-			display: flex;
-			align-items: center;
-			justify-content: center;
-			height: 14px;
-			width: 14px;
-			background-color: rgba(192, 193, 195, 0.2); /* #C0C1C3 with 0.2 opacity */
-			border-radius: 2px;
-		}
-	}
-}
