+++ frontend/src/types/api/integrations/aws.ts
@@ -1,4 +1,4 @@
-import { CloudAccount } from 'pages/Integrations/CloudIntegrationPage/ServicesSection/types';
+import { CloudAccount } from 'container/CloudIntegrationPage/ServicesSection/types';
 
 export interface GenerateConnectionUrlPayload {
 	agent_config: {
